refactor(main page): #Y-12 Сверстать главную страницу игры: меню. Скорректировать код после ревью
