+++ packages/client/src/pages/mainPage/index.tsx
@@ -1,15 +1,16 @@
 import styles from './mainPage.module.css'
 import main from '../../assets/images/main.png'
-import { Menu } from '../../components/menu/Menu'
+import { MAIN_MENU_LINKS } from '../../utils/constants'
+import { Menu } from '../../components'
 
 export const MainPage = () => {
   return (
     <section className={styles.main}>
       <div className={styles.content}>
         <h1 className={styles.title}>CHESS TANKS</h1>
-        <Menu />
+        <Menu title="Основное меню" links={MAIN_MENU_LINKS} />
       </div>
-      <img src={main} alt="Танки" />
+      <img src={main} alt="Танки" className={styles.image} />
     </section>
   )
 }
