+++ packages/client/src/utils/constants.ts
@@ -1,0 +1,6 @@
+export const MAIN_MENU_LINKS = [
+  { to: '/game', label: 'Начать игру' },
+  { to: '/profile', label: 'Профиль' },
+  { to: '/rating', label: 'Лидербоард' },
+  { to: '/forum', label: 'Форум' },
+]
